% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{epiverse_conflicts}
\alias{epiverse_conflicts}
\title{Conflicts between epiverse packages and other packages}
\usage{
epiverse_conflicts(only = NULL)
}
\arguments{
\item{only}{Set this to a character vector to restrict to conflicts only
with these packages.}
}
\value{
An object of class \code{epiverse_conflicts}, invisibly. Printed output
shows which epiverse functions mask functions from other packages.
}
\description{
Lists all function conflicts between epiverse packages and other packages
that you have loaded.
}
\examples{
\dontrun{
epiverse_conflicts()
}
}
