% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiverse-package.R
\docType{package}
\name{epiverse-package}
\alias{epiverse}
\alias{epiverse-package}
\title{epiverse: Easily Install and Load the EPI R Package Suite}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'epiverse' package makes it easy to install and load the collection of R packages maintained by the Economic Policy Institute. Loading 'epiverse' attaches all included packages and reports any function conflicts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://economic.github.io/epiverse/}
  \item \url{https://github.com/Economic/epiverse}
  \item Report bugs at \url{https://github.com/economic/epiverse/issues}
}

}
\author{
\strong{Maintainer}: Ben Zipperer \email{benzipperer@gmail.com}

Other contributors:
\itemize{
  \item Economic Policy Institute [copyright holder, funder]
}

}
\keyword{internal}
