% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epiverse-package.R
\docType{package}
\name{Epiverse-package}
\alias{Epiverse}
\alias{Epiverse-package}
\title{Epiverse: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emjnixon15/Epiverse}
  \item Report bugs at \url{https://github.com/emjnixon15/Epiverse/issues}
}

}
\author{
\strong{Maintainer}: Emily Nixon \email{emily.nixon@liverpool.ac.uk} (\href{https://orcid.org/0000-0002-1626-9296}{ORCID}) [copyright holder]

}
\keyword{internal}
