% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{epiverse-package}
\alias{epiverse}
\alias{epiverse-package}
\title{epiverse: Meta-Package for the Epiverse-TRACE Project}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A meta-package to provide easy access to all 'Epiverse-TRACE' packages released on CRAN. From a technical perspective, it is also where integration tests across packages are hosted, and where dependency scaffolding at the level of this ecosystem is investigated.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/epiverse}
  \item \url{https://epiverse-trace.github.io/epiverse}
  \item Report bugs at \url{https://github.com/epiverse-trace/epiverse/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})

}
\keyword{internal}
